moves posture data to not time out
