+++ controller/model/posture_response_model.go
@@ -32,7 +32,6 @@ const (
 type PostureCache struct {
 	identityToPostureData cmap.ConcurrentMap //identityId -> PostureData
 	ticker                *time.Ticker
-	postureDataTimeout    time.Duration
 	events.EventEmmiter
 }
 
@@ -41,41 +40,9 @@ func newPostureCache() *PostureCache {
 		identityToPostureData: cmap.New(),
 		ticker:                time.NewTicker(10 * time.Second),
 		EventEmmiter:          events.New(),
-		postureDataTimeout:    30 * time.Second,
 	}
 
-	pc.start()
-
 	return pc
-}
-
-func (pc *PostureCache) start() {
-	go func() {
-		for range pc.ticker.C {
-
-			changedIdentityIds := map[string]struct{}{}
-
-			for _, identityId := range pc.identityToPostureData.Keys() {
-				pc.identityToPostureData.Upsert(identityId, newPostureData(), func(exist bool, valueInMap interface{}, newValue interface{}) interface{} {
-					var postureData *PostureData
-					if exist {
-						postureData = valueInMap.(*PostureData)
-					} else {
-						postureData = newPostureData()
-					}
-
-					if changed := postureData.Timeout(time.Now().Add(-1 * pc.postureDataTimeout)); changed {
-						changedIdentityIds[identityId] = struct{}{}
-					}
-
-					return postureData
-				})
-			}
-			for identityId := range changedIdentityIds {
-				pc.Emit(EventIdentityPostureDataAltered, identityId)
-			}
-		}
-	}()
 }
 
 func (pc *PostureCache) Add(identityId string, postureResponses []*PostureResponse) {
@@ -125,19 +92,6 @@ type PostureData struct {
 	ApiSessions map[string]*ApiSessionPostureData `json:"sessionPostureData"`
 }
 
-func (pd *PostureData) Timeout(oldest time.Time) bool {
-	changed := false
-	changed = pd.Mac.Timeout(oldest) || changed
-	changed = pd.Domain.Timeout(oldest) || changed
-	changed = pd.Os.Timeout(oldest) || changed
-
-	for _, process := range pd.Processes {
-		changed = process.Timeout(oldest) || changed
-	}
-
-	return changed
-}
-
 func (pd *PostureData) Evaluate(apiSessionId string, checks []*PostureCheck) bool {
 	for _, check := range checks {
 		if result := check.Evaluate(apiSessionId, pd); !result {
@@ -177,128 +131,16 @@ type PostureResponse struct {
 	SubType        PostureResponseSubType `json:"-"`
 }
 
-func (pr *PostureResponse) Timeout(oldest time.Time) bool {
-	if !pr.TimedOut && pr.LastUpdatedAt.Before(oldest) {
-		pr.TimedOut = true
-		return true
-	}
-
-	return false
-}
-
 func (pr *PostureResponse) Apply(postureData *PostureData) {
 	pr.SubType.Apply(postureData)
 }
 
 type PostureResponseSubType interface {
 	Apply(postureData *PostureData)
-	Timeout(oldest time.Time) bool
-}
-
-type PostureResponseMac struct {
-	*PostureResponse
-	Addresses []string `json:"addresses"`
-}
-
-func (pr *PostureResponseMac) Apply(postureData *PostureData) {
-	var cleanedAddresses []string
-	for _, address := range pr.Addresses {
-		cleanedAddresses = append(cleanedAddresses, CleanHexString(address))
-	}
-
-	pr.Addresses = cleanedAddresses
-
-	postureData.Mac = pr
-	postureData.Mac.LastUpdatedAt = time.Now()
 }
 
 var macClean = regexp.MustCompile("[^a-f0-9]+")
 
 func CleanHexString(hexString string) string {
 	return macClean.ReplaceAllString(strings.ToLower(hexString), "")
-}
-
-type PostureResponseOs struct {
-	*PostureResponse
-	Type    string `json:"type"`
-	Version string `json:"version"`
-	Build   string `json:"build"`
-}
-
-func (pr *PostureResponseOs) Apply(postureData *PostureData) {
-	postureData.Os = pr
-	postureData.Os.LastUpdatedAt = time.Now()
-}
-
-type PostureResponseProcess struct {
-	*PostureResponse
-	IsRunning          bool     `json:"isRunning"`
-	BinaryHash         string   `json:"binaryHash"`
-	SignerFingerprints []string `json:"signerFingerprints"`
-}
-
-func (pr *PostureResponseProcess) Apply(postureData *PostureData) {
-	found := false
-
-	for i, fingerprint := range pr.SignerFingerprints {
-		pr.SignerFingerprints[i] = CleanHexString(fingerprint)
-	}
-
-	pr.BinaryHash = CleanHexString(pr.BinaryHash)
-
-	for i, process := range postureData.Processes {
-		if process.PostureCheckId == pr.PostureCheckId {
-			postureData.Processes[i] = pr
-			postureData.Processes[i].LastUpdatedAt = time.Now()
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		pr.LastUpdatedAt = time.Now()
-		postureData.Processes = append(postureData.Processes, pr)
-	}
-}
-
-type PostureResponseDomain struct {
-	*PostureResponse
-	Name string `json:"name"`
-}
-
-func (pr *PostureResponseDomain) Apply(postureData *PostureData) {
-	postureData.Domain = pr
-	postureData.Domain.LastUpdatedAt = time.Now()
-}
-
-const (
-	MfaProviderZiti string = "ziti"
-)
-
-type PostureResponseMfa struct {
-	*PostureResponse
-	ApiSessionId string `json:"-"'`
-	PassedMfa    bool   `json:"passedMfa"`
-}
-
-func (pr *PostureResponseMfa) Apply(postureData *PostureData) {
-
-	if postureData.ApiSessions == nil {
-		postureData.ApiSessions = map[string]*ApiSessionPostureData{}
-	}
-
-	apiSessionPostureData := postureData.ApiSessions[pr.ApiSessionId]
-
-	if apiSessionPostureData == nil {
-		apiSessionPostureData = &ApiSessionPostureData{}
-		postureData.ApiSessions[pr.ApiSessionId] = apiSessionPostureData
-	}
-
-	apiSessionPostureData.Mfa = pr
-	apiSessionPostureData.Mfa.LastUpdatedAt = time.Now()
-}
-
-func (pr *PostureResponseMfa) Timeout(oldest time.Time) bool {
-	//valid for lifetime of ApiSession
-	return false
 }
