+++ controller/model/posture_response_model_domain.go
@@ -1,0 +1,31 @@
+/*
+	Copyright NetFoundry, Inc.
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+	https://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+*/
+
+package model
+
+import (
+	"time"
+)
+
+type PostureResponseDomain struct {
+	*PostureResponse
+	Name string `json:"name"`
+}
+
+func (pr *PostureResponseDomain) Apply(postureData *PostureData) {
+	postureData.Domain = pr
+	postureData.Domain.LastUpdatedAt = time.Now()
+}
