+++ controller/model/posture_response_model_mfa.go
@@ -1,0 +1,46 @@
+/*
+	Copyright NetFoundry, Inc.
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+	https://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+*/
+
+package model
+
+import "time"
+
+const (
+	MfaProviderZiti string = "ziti"
+)
+
+type PostureResponseMfa struct {
+	*PostureResponse
+	ApiSessionId string `json:"-"`
+	PassedMfa    bool   `json:"passedMfa"`
+}
+
+func (pr *PostureResponseMfa) Apply(postureData *PostureData) {
+
+	if postureData.ApiSessions == nil {
+		postureData.ApiSessions = map[string]*ApiSessionPostureData{}
+	}
+
+	apiSessionPostureData := postureData.ApiSessions[pr.ApiSessionId]
+
+	if apiSessionPostureData == nil {
+		apiSessionPostureData = &ApiSessionPostureData{}
+		postureData.ApiSessions[pr.ApiSessionId] = apiSessionPostureData
+	}
+
+	apiSessionPostureData.Mfa = pr
+	apiSessionPostureData.Mfa.LastUpdatedAt = time.Now()
+}
