+++ controller/model/posture_response_model_process.go
@@ -1,0 +1,52 @@
+/*
+	Copyright NetFoundry, Inc.
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+	https://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+*/
+
+package model
+
+import (
+	"time"
+)
+
+type PostureResponseProcess struct {
+	*PostureResponse
+	IsRunning          bool     `json:"isRunning"`
+	BinaryHash         string   `json:"binaryHash"`
+	SignerFingerprints []string `json:"signerFingerprints"`
+}
+
+func (pr *PostureResponseProcess) Apply(postureData *PostureData) {
+	found := false
+
+	for i, fingerprint := range pr.SignerFingerprints {
+		pr.SignerFingerprints[i] = CleanHexString(fingerprint)
+	}
+
+	pr.BinaryHash = CleanHexString(pr.BinaryHash)
+
+	for i, process := range postureData.Processes {
+		if process.PostureCheckId == pr.PostureCheckId {
+			postureData.Processes[i] = pr
+			postureData.Processes[i].LastUpdatedAt = time.Now()
+			found = true
+			break
+		}
+	}
+
+	if !found {
+		pr.LastUpdatedAt = time.Now()
+		postureData.Processes = append(postureData.Processes, pr)
+	}
+}
