+++ specs/swagger.yml
@@ -5772,6 +5772,9 @@ definitions:
     properties:
       args:
         $ref: '#/definitions/apiErrorArgs'
+      data:
+        type: object
+        additionalProperties: true
       causeMessage:
         type: string
       code:
